explain: move some structs from `adapter` to `repr`
Move the following `explain_new` structs from `mz_adapter` to `mz_repr`:

- `AnnotatedPlan`,
- `Attributes`,
- `UsedIndexes`.
